pyproject.toml
[project]
name = "dual-source-news-scraper"
version = "1.0.0"
description = "A dual-source financial news scraper for CNBC and CNN"
authors = ["Developer"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12.7"
httpx = "^0.25.0"
beautifulsoup4 = "^4.12.0"
python-dateutil = "^2.8.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"